re-add check user

+++ src/bot.py
@@ -69,6 +69,12 @@ class MyCog(commands.Cog):
 
             temp_user = self.bot.get_user(user_id)
 
+            if temp_user is None:
+                print(f"USER {TEMP_ID} NOT FOUND")
+                return
+
+            print(f"{temp_user.name} was found!")
+
             print(message.body)
             await temp_user.send(message.body)
             message.delete()
